update for python3

+++ main.py
@@ -1,15 +1,27 @@
 # coding=utf-8
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import sys
+
+PY3 = True
+if sys.version_info.major == 2:
+    PY3 = False
+    reload(sys)
+    sys.setdefaultencoding("utf-8")
+    
+    from ConfigParser import ConfigParser
+    import commands
+    from urllib2 import urlopen
+else:
+    from configparser import ConfigParser
+    import subprocess as commands
+    from urllib.request import urlopen
 
 import os
-import sys
 import io
-from ConfigParser import ConfigParser
 import plistlib
-import urllib
 import json
-import commands
-reload(sys)
-sys.setdefaultencoding("utf-8")
 
 import rumps
 from genpac import update_gfwlist
@@ -37,17 +49,17 @@ pac_port=1082
 '''
 
 APP_MENU = [
-    u"当前状态:关闭",
-    u"✅启动",
+    "当前状态:关闭",
+    "✅启动",
     None,
-    u"PAC模式",
-    u"全局模式",
+    "PAC模式",
+    "全局模式",
     None,
-    u"⚙设置",
-    u"更新trojan core",
-    u"从gfwlist更新PAC",
+    "⚙设置",
+    "更新trojan core",
+    "从gfwlist更新PAC",
     None,
-    u"退出",
+    "退出",
 ]
 
 CONFIG_DIR = os.path.join(os.environ.get("HOME", "~"), ".TrojanA")
@@ -57,6 +69,33 @@ CONFIG_PATH = os.path.join(CONFIG_DIR, "app.ini")
 APP_CONFIG = {}
 
 TROJAN_CONFIG = {}
+
+
+def plist_write(data, file_name):
+    """
+    保存plist文件
+    """
+    if not PY3:
+        plistlib.writePlist(data, file_name)
+        return
+    with open(file_name, "wb") as f:
+        plistlib.dump(data, f)
+
+
+def file_read(file_name):
+    """
+    读取文件内容
+    """
+    with io.open(file_name, encoding="utf-8") as f:
+        return f.read()
+
+
+def file_write(data, file_name):
+    """
+    将数据写入文件
+    """
+    with io.open(file_name, mode="w", encoding="utf-8") as f:
+        f.write(data)
 
 
 def gen_http_plist():
@@ -72,7 +111,7 @@ def gen_http_plist():
         'WorkingDirectory': os.path.join(CONFIG_DIR, "pac"),
         'StandardOutPath': os.path.join(home_dir, 'Library/Logs/trojana-http.log'),
         'Label': 'com.chrisxiao.trojana.http-server',
-        'ProgramArguments': ['python', os.path.abspath('./httpserver.py'), APP_CONFIG.get("pac_port", "1082")],
+        'ProgramArguments': [sys.executable, os.path.abspath('./httpserver.py'), APP_CONFIG.get("pac_port", "1082")],
         'KeepAlive': True,
         'StandardErrorPath': os.path.join(home_dir, 'Library/Logs/trojana-http.log')
     }
@@ -81,7 +120,7 @@ def gen_http_plist():
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", http_plist)
 
     # print plist_file
-    plistlib.writePlist(data, plist_file)
+    plist_write(data, plist_file)
 
 
 def gen_trojan_plist():
@@ -106,7 +145,7 @@ def gen_trojan_plist():
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", trojan_plist)
 
     # print plist_file
-    plistlib.writePlist(data, plist_file)
+    plist_write(data, plist_file)
 
 
 def init_config():
@@ -122,10 +161,8 @@ def init_config():
     if not os.path.exists(pac_dir):
         os.mkdir(pac_dir)
 
-    if not os.path.exists(CONFIG_PATH):
-        fp = file(CONFIG_PATH, mode="w")
-        fp.write(CONFIG_TEMPLATE)
-        fp.close()
+    if not os.path.exists(CONFIG_PATH): 
+        file_write(CONFIG_TEMPLATE, CONFIG_PATH)
 
 
 def load_config():
@@ -133,7 +170,9 @@ def load_config():
     load config
     """
     config = ConfigParser()
-    config.read(CONFIG_PATH)
+    with io.open(CONFIG_PATH, encoding="utf-8") as f:
+        config.readfp(f)
+
     APP_CONFIG.update(dict(config.items("App")))
     TROJAN_CONFIG.update(dict(config.items("Trojan")))
 
@@ -150,37 +189,35 @@ def flush_config(config_string):
     """
     flush config
     """
-    
-    config = ConfigParser()
-    config.readfp(io.BytesIO(config_string))
 
+    config = ConfigParser()
+    config.readfp(io.StringIO(config_string))
     APP_CONFIG.update(dict(config.items("App")))
     TROJAN_CONFIG.update(dict(config.items("Trojan")))
 
-    file(CONFIG_PATH, "w").write(config_string)
-    print "flush config ok"
+    file_write(config_string, CONFIG_PATH)
+    print("flush config ok")
 
     render_pac("./pac/gfwlist.js.tpl", os.path.join(CONFIG_DIR, "pac", "gfwlist.js"))
     render_trojan_conf("./trojan/config.json.tpl", os.path.join(CONFIG_DIR, "config.json"))
-    print "rewrite config file ok"
+    print("rewrite config file ok")
 
 
 def render_pac(tpl_file, dst_file):
     """
     根据配置信息生成pac文件
     """
-    tpl = file(tpl_file).read()
+    tpl = file_read(tpl_file)
     tpl = tpl.replace("__PROXY_PORT__", TROJAN_CONFIG["local_port"])
     # print tpl
-    file(dst_file, "w").write(tpl)
+    file_write(tpl, dst_file)
 
 
 def render_trojan_conf(tpl_file, dst_file):
     """
     根据配置信息生成trojan客户端配置文件
     """
-
-    tpl = file(tpl_file).read()
+    tpl = file_read(tpl_file)
 
     tpl = tpl.replace("__LOCAL_PORT__", TROJAN_CONFIG["local_port"])
     tpl = tpl.replace("__REMOTE_ADDR__", TROJAN_CONFIG["remote_addr"])
@@ -190,14 +227,14 @@ def render_trojan_conf(tpl_file, dst_file):
     tpl = tpl.replace("__VERIFY_HOSTNAME__", TROJAN_CONFIG["ssl.verify_hostname"])
     tpl = tpl.replace("__CERT__", TROJAN_CONFIG["ssl.cert"])
     # print tpl
-    file(dst_file, "w").write(tpl)
+    file_write(tpl, dst_file)
 
 
 def pac_mode_on(pac_url):
     """
     打开pac模式
     """
-    print "pac on"
+    print("pac on")
     os.system("networksetup -setautoproxyurl 'Wi-Fi' %s" % pac_url)
 
 
@@ -205,7 +242,7 @@ def pac_mode_off():
     """
     关闭pac模式
     """
-    print "pac off"
+    print("pac off")
     os.system("networksetup -setautoproxystate 'Wi-Fi' 'off'")
 
 
@@ -213,7 +250,7 @@ def global_mode_on(host, port):
     """
     打开全局代理模式
     """
-    print "global on"
+    print("global on")
     os.system("networksetup -setsocksfirewallproxy 'Wi-Fi' %s %s" % (host, port))
 
 
@@ -221,7 +258,7 @@ def global_mode_off():
     """
     关闭全局代理模式
     """
-    print "global off"
+    print("global off")
     os.system("networksetup -setsocksfirewallproxystate 'Wi-Fi' off")
 
 
@@ -233,8 +270,8 @@ def start_trojan():
     trojan_plist = "com.chrisxiao.trojana.trojan-client.plist"
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", trojan_plist)
 
-    print "trojan started"
-    print os.system("launchctl load " + plist_file)
+    print("trojan started")
+    print(os.system("launchctl load " + plist_file))
 
 
 def stop_trojan():
@@ -245,8 +282,8 @@ def stop_trojan():
     trojan_plist = "com.chrisxiao.trojana.trojan-client.plist"
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", trojan_plist)
 
-    print "trojan stop"
-    print os.system("launchctl unload " + plist_file)
+    print("trojan stop")
+    print(os.system("launchctl unload " + plist_file))
 
 
 def start_pac_server():
@@ -257,8 +294,8 @@ def start_pac_server():
     http_plist = "com.chrisxiao.trojana.http-server.plist"
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", http_plist)
 
-    print "pac server started"
-    print os.system("launchctl load " + plist_file)
+    print("pac server started")
+    print(os.system("launchctl load " + plist_file))
 
 
 def stop_pac_server():
@@ -269,8 +306,8 @@ def stop_pac_server():
     http_plist = "com.chrisxiao.trojana.http-server.plist"
     plist_file = os.path.join(home_dir, "Library/LaunchAgents", http_plist)
 
-    print "pac server stop"
-    print os.system("launchctl unload " + plist_file)
+    print("pac server stop")
+    print(os.system("launchctl unload " + plist_file))
 
 
 def get_cur_trojan_ver():
@@ -286,7 +323,7 @@ def get_latest_trojan():
     获取最新trojan版本信息
     """
 
-    r = urllib.urlopen("https://api.github.com/repos/trojan-gfw/trojan/releases/latest")
+    r = urlopen("https://api.github.com/repos/trojan-gfw/trojan/releases/latest", timeout=10)
     res_json = json.loads(r.read())
 
     release_ver = res_json["name"].split()[-1]
@@ -300,26 +337,32 @@ def do_update_trojan():
     检查并更新trojan版本
     """
     now_ver = get_cur_trojan_ver()
-    latest_info = get_latest_trojan()
+    try:
+        latest_info = get_latest_trojan()
+    except Exception as e:
+        print(e)
+        return False, "无法获取最新版本信息！"
 
     if now_ver >= latest_info[0]:
-        return False
+        return False, "已是最新版本，无需更新！"
 
-    print "start downloading new"
+    print("start downloading new")
     url = latest_info[1][0]
     target = "./download/" + os.path.basename(latest_info[1][0])
-    print target
+    print(target)
 
-    os.system("curl -sL %s --output %s" % (url, target))
+    ret = os.system("curl --connect-timeout 10 --max-time 60 -L %s --output %s" % (url, target))
+    if ret != 0:
+        return False, "无法下载最新版本"
     os.system("unzip -u %s -d ./download/tmp" % target)
     os.system("rm -f ./trojan/trojan")
     os.system("cp ./download/tmp/trojan/trojan ./trojan/trojan")
     os.system("rm -rf ./download/tmp")
     os.system("rm -f %s" % target)
 
-    print "update finished"
+    print("update finished")
 
-    return True
+    return True, "已是最新版本，无需更新！"
 
 
 
@@ -329,11 +372,11 @@ class AwesomeStatusBarApp(rumps.App):
     pac_item = None
     global_item = None
     status_item = None
-    config_window = rumps.Window(u"",
-                                 title=u"基本配置",
+    config_window = rumps.Window("",
+                                 title="基本配置",
                                  default_text=CONFIG_TEMPLATE,
-                                 ok=u"确定",
-                                 cancel=u"取消",
+                                 ok="确定",
+                                 cancel="取消",
                                  dimensions=(480, 320)
                                 )
     pac_url = "http://127.0.0.1:%s/gfwlist.js"
@@ -342,9 +385,9 @@ class AwesomeStatusBarApp(rumps.App):
         """
         初始化
         """
-        self.pac_item = self.menu[u"PAC模式"]
+        self.pac_item = self.menu["PAC模式"]
         self.pac_item.state = True
-        self.global_item = self.menu[u"全局模式"]
+        self.global_item = self.menu["全局模式"]
         self.global_item.state = False
 
         # start_trojan()
@@ -353,17 +396,17 @@ class AwesomeStatusBarApp(rumps.App):
         # pac_mode_on(self.pac_url)
 
 
-    @rumps.clicked(u"✅启动")
+    @rumps.clicked("✅启动")
     def onoff(self, sender):
         if not self.status_item:
-            self.status_item = self._menu[u"当前状态:关闭"]
+            self.status_item = self._menu["当前状态:关闭"]
 
-        self.status_item.title = u"当前状态:启动"
+        self.status_item.title = "当前状态:启动"
 
-        sender.title = u"❌关闭" if sender.title == u"✅启动" else u"✅启动"
-        self.status_item.title = u"当前状态:关闭" if sender.title == u"✅启动" else u"当前状态:开启"
+        sender.title = "❌关闭" if sender.title == "✅启动" else "✅启动"
+        self.status_item.title = "当前状态:关闭" if sender.title == "✅启动" else "当前状态:开启"
 
-        if self.status_item.title == u"当前状态:关闭":
+        if self.status_item.title == "当前状态:关闭":
             stop_trojan()
             global_mode_off()
             pac_mode_off()
@@ -380,17 +423,17 @@ class AwesomeStatusBarApp(rumps.App):
         rumps.alert(self.status_item.title)
 
 
-    @rumps.clicked(u"⚙设置")
+    @rumps.clicked("⚙设置")
     def settings(self, _):
-        self.config_window.default_text = file(CONFIG_PATH).read()
+        self.config_window.default_text = file_read(CONFIG_PATH)
         res = self.config_window.run()
         if res.clicked:
-            print res.clicked
+            print(res.clicked)
             # print res.text
-            flush_config(str(res.text))
+            flush_config(res.text)
 
 
-    @rumps.clicked(u"PAC模式")
+    @rumps.clicked("PAC模式")
     def pac_mode(self, sender):
         sender.state = True
         self.global_item.state = False
@@ -400,7 +443,7 @@ class AwesomeStatusBarApp(rumps.App):
         pac_mode_on(self.pac_url)
 
 
-    @rumps.clicked(u"全局模式")
+    @rumps.clicked("全局模式")
     def global_mode(self, sender):
         sender.state = True
         self.pac_item.state = False
@@ -410,26 +453,25 @@ class AwesomeStatusBarApp(rumps.App):
         global_mode_on("127.0.0.1", TROJAN_CONFIG["local_port"])
 
 
-    @rumps.clicked(u"更新trojan core")
+    @rumps.clicked("更新trojan core")
     def update_trojan_client(self, sender):
-        if do_update_trojan():
-            if hasattr(self.status_item, "title") and self.status_item.title == u"当前状态:开启":
+        result, message = do_update_trojan()
+        if result:
+            if hasattr(self.status_item, "title") and self.status_item.title == "当前状态:开启":
                 stop_trojan()
                 start_trojan()
-            rumps.alert(u"更新成功!")
-        else:
-            rumps.alert(u"已是最新版本，无需更新！")
+        rumps.alert(message)
 
 
-    @rumps.clicked(u"从gfwlist更新PAC")
+    @rumps.clicked("从gfwlist更新PAC")
     def update_pac(self, sender):
         if update_gfwlist(os.path.join(CONFIG_DIR, "pac", "gfwlist.js"), TROJAN_CONFIG["local_port"]):
-            rumps.alert(u"更新成功!")
+            rumps.alert("更新成功!")
         else:
-            rumps.alert(u"已是最新版本，无需更新！")
+            rumps.alert("已是最新版本，无需更新！")
 
 
-    @rumps.clicked(u"退出")
+    @rumps.clicked("退出")
     def quit(self, _):
         global_mode_off()
         pac_mode_off()
